The bin directory contains the ".exe" documents that cane executed. And the remain documents are all C code documents.

+++ C-Code/excerice/area2.c
@@ -1,7 +1,11 @@
+/**
+ * this is a algorithm that can do "(x + y) * z / 2" work.
+*/
+
 #include <stdio.h>
 #include <math.h>
 
-float main(){
+int main(){
     float x, y, z;
     scanf("%f,%f,%f",&x,&y,&z);
     printf("%.2f\n", (x + y) * z / 2);
